refactor: rename @kubricate/env to @kubricate/plugin-env across the codebase
