+++ README.md
@@ -381,7 +381,7 @@ Documentation is in progress, please explore the [`examples`](https://github.com
 
 - `kubricate` – CLI for configuration and manifest generation
 - `@kubricate/core` – Core framework for creating and managing stacks
-- `@kubricate/env` – Secret connector for `.env` and environment variables
+- `@kubricate/plugin-env` – Secret connector for `.env` and environment variables
 - `@kubricate/kubernetes` – Kubernetes connectors
 - `@kubricate/stacks` – Official reusable stack definitions
 - `@kubricate/toolkit` – Utility functions for custom stack authors
@@ -392,7 +392,7 @@ Ensure the following packages are always on the same version when upgrading:
 
 - `kubricate`
 - `@kubricate/core`
-- `@kubricate/env`
+- `@kubricate/plugin-env`
 - `@kubricate/kubernetes`
 - `@kubricate/stacks`
