+++ examples/with-secret-registry/src/setup-secret.ts
@@ -1,6 +1,6 @@
 import { SecretManager, SecretRegistry } from '@kubricate/core';
 import { OpaqueSecretProvider } from '@kubricate/kubernetes';
-import { EnvConnector } from '@kubricate/env';
+import { EnvConnector } from '@kubricate/plugin-env';
 
 const frontendSecretManager = new SecretManager()
   .addConnector('EnvConnector', new EnvConnector())
