+++ tests/fixtures/shared-configs.ts
@@ -1,7 +1,7 @@
 import { NamespaceStack, SimpleAppStack } from "@kubricate/stacks";
 import { SecretManager, SecretRegistry } from '@kubricate/core';
 import { OpaqueSecretProvider, DockerConfigSecretProvider } from '@kubricate/kubernetes';
-import { EnvConnector } from '@kubricate/env';
+import { EnvConnector } from '@kubricate/plugin-env';
 
 export const frontendSecretManager = new SecretManager()
   .addConnector('EnvConnector', new EnvConnector())
